#52 Remove SQlite3 from installation folder on Linux

+++ resources/build-local.sh
@@ -1,10 +1,17 @@
+if [ $# -eq 0 ]
+  then
+    echo "No arguments supplied, 'drX' expected"
+    exit 1
+fi
+
 game=$1
 echo "Building for $game"
 rm -f dirtrally-lb.db
-./bin/sqlite3.exe -batch -init resources/setup-${game}.sql dirtrally-lb.db .exit 
+sqlite3 -batch -init resources/setup-${game}.sql dirtrally-lb.db .exit 
 pyinstaller timerecord.spec --noconfirm --name timerecorder-${game}
 
 cd dist/timerecorder-${game}
+rm sqlite3.exe
 mv list-laptimes.bat list-laptimes.sh
 mv export-laptimes.bat export-laptimes.sh
 # Reduce scripts to core parts
